[FIX] bookstore_mgmt_google_books_api: Fixed book importation for product.product
Refactored to use a mixin for Google Books API integration. This mixin can now be inherited by both product.template and product.product,  enabling proper functionality for both models.
