+++ bookstore_mgmt_google_books_api/README.rst
@@ -69,10 +69,11 @@ Authors
 Contributors
 ------------
 
-- [APSL-Nagarro](https://apsl.tech):
+-  [APSL-Nagarro](https://apsl.tech):
 
-  - Miquel Alzanillas <malzanillas@apsl.net>
-  - Antoni Marroig <amarroig@apsl.net>
+   -  Miquel Alzanillas <malzanillas@apsl.net>
+   -  Antoni Marroig <amarroig@apsl.net>
+   -  Bernat Obrador <bobrador@apsl.net>
 
 Maintainers
 -----------
