+++ bookstore_mgmt_google_books_api/i18n/bookstore_mgmt_google_books_api.pot
@@ -6,6 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 17.0\n"
 "Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2025-01-16 10:09+0000\n"
+"PO-Revision-Date: 2025-01-16 10:09+0000\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -15,16 +17,14 @@ msgstr ""
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "Book data updated from Google Books API"
 msgstr ""
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "ISBN code is mandatory. Please, provide one."
 msgstr ""
@@ -37,8 +37,7 @@ msgstr ""
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "Not book found with this data"
 msgstr ""
@@ -55,9 +54,13 @@ msgstr ""
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #: code:addons/bookstore_mgmt_google_books_api/tests/test_google_books_api.py:0
 #, python-format
 msgid "Warning"
+msgstr ""
+
+#. module: bookstore_mgmt_google_books_api
+#: model:ir.model,name:bookstore_mgmt_google_books_api.model_google_books_mixin
+msgid "google.books.mixin"
 msgstr ""
