+++ bookstore_mgmt_google_books_api/i18n/es.po
@@ -6,11 +6,10 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 17.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-11-13 15:45+0000\n"
-"PO-Revision-Date: 2024-11-13 15:45+0000\n"
+"POT-Creation-Date: 2025-01-16 10:10+0000\n"
+"PO-Revision-Date: 2025-01-16 10:10+0000\n"
 "Last-Translator: \n"
 "Language-Team: \n"
-"Language: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
@@ -18,16 +17,14 @@ msgstr ""
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "Book data updated from Google Books API"
 msgstr "Datos actualizados desde la API de Google Libros"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "ISBN code is mandatory. Please, provide one."
 msgstr "El código ISBN es obligatorio. Por favor, proporcione uno."
@@ -40,8 +37,7 @@ msgstr "Importar datos del libro"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "Not book found with this data"
 msgstr "No se ha encontrado ningún libro con estos datos"
@@ -58,9 +54,13 @@ msgstr "Variante de producto"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #: code:addons/bookstore_mgmt_google_books_api/tests/test_google_books_api.py:0
 #, python-format
 msgid "Warning"
 msgstr "Advertencia"
+
+#. module: bookstore_mgmt_google_books_api
+#: model:ir.model,name:bookstore_mgmt_google_books_api.model_google_books_mixin
+msgid "google.books.mixin"
+msgstr ""
