+++ bookstore_mgmt_google_books_api/i18n/it.po
@@ -6,28 +6,25 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 17.0\n"
 "Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2024-11-25 11:06+0000\n"
-"Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
-"Language-Team: none\n"
-"Language: it\n"
+"POT-Creation-Date: 2025-01-16 10:10+0000\n"
+"PO-Revision-Date: 2025-01-16 10:10+0000\n"
+"Last-Translator: \n"
+"Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 5.6.2\n"
+"Plural-Forms: \n"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "Book data updated from Google Books API"
 msgstr "Dati libro aggiornati da API Google Books"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "ISBN code is mandatory. Please, provide one."
 msgstr "Il codice ISBN è obbligatorio. Inserirne uno."
@@ -40,8 +37,7 @@ msgstr "Importa dati libro"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #, python-format
 msgid "Not book found with this data"
 msgstr "Nessun libro trovato con questi dati"
@@ -58,9 +54,13 @@ msgstr "Variante prodotto"
 
 #. module: bookstore_mgmt_google_books_api
 #. odoo-python
-#: code:addons/bookstore_mgmt_google_books_api/models/product_product.py:0
-#: code:addons/bookstore_mgmt_google_books_api/models/product_template.py:0
+#: code:addons/bookstore_mgmt_google_books_api/models/google_books_mixin.py:0
 #: code:addons/bookstore_mgmt_google_books_api/tests/test_google_books_api.py:0
 #, python-format
 msgid "Warning"
 msgstr "Attenzione"
+
+#. module: bookstore_mgmt_google_books_api
+#: model:ir.model,name:bookstore_mgmt_google_books_api.model_google_books_mixin
+msgid "google.books.mixin"
+msgstr ""
