+++ bookstore_mgmt_google_books_api/models/__init__.py
@@ -1,4 +1,5 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
+from . import google_books_mixin
 from . import product_product
 from . import product_template
